Use a lattice representation of PermissionPresenceSet
Represent PermissionPresence as a product lattice of Usage and Capability.

* Usage is a simple two element lattice - either we know nothing about how a
  permission name is used, or we know that it's used by a call that has a 'Uses'
  action.

* Capability is a 4 element lattice that tracks what we know about a permission
  name at each site in a call sequence: CapUnknown (bottom) - we don't know anything,
  CapHas - the site has the capability, CapLacks - the site lacks
  the capablity, CapConflict - we learned conflicting information (both
  'CapHas' and 'CapLacks') along different paths in the call tree.

* PermissionPresence is a simple product lattice: its bottom is the pair bottom
  bottom and its join and partial order are given component-wise.

The PermissionPresence lattice extends to a lattice on
PermissionPresenceSet (now a finite map from PermissionName to
PermissionPresence): the empty set is the bottom element, and the join of two
finite maps takes their union applying the join operation of PermissionPresence
when the same name is present in both maps.

With this, we can see that most of the propagation during inference is simple
joins of old and new information except when propagating forward through 'Call'
site which has a 'Grant' or 'Revoke' action which overwrites the Capability
portion of the prior information (which are in some sense stale from the
previous iteration).